Move rendering related functions to render module
